+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ### Added | Changed | Deprecated | Removed | Fixed | Security -->
 
+### Fixed
+
+- Prevent duplicate descendants
+
 ### Changed
 
 - Upgraded k8s dependency to 2.0.x. [#173](https://github.com/coryodaniel/bonny/pull/173)
@@ -21,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Helpers to define conditions on custom resources. [#188](https://github.com/coryodaniel/bonny/pull/188)
   - `Bonny.API.Version.add_conditions/1` - Helper to define a schema for conditions on the CRD
   - `Bonny.Axn.set_condition/4` - Helper to set a condition on a resource
+- `Bonny.Pluggable.Finalizer` - A pluggable step to define and implement finalizers
 
 <!-- No new entries below this line! -->